+++ lizardwizards/src/main/java/com/lizardwizards/lizardwizards/client/LizardWizardsApplication.java
@@ -1,6 +1,5 @@
 package com.lizardwizards.lizardwizards.client;
 
-import com.lizardwizards.lizardwizards.core.Utils;
 import javafx.application.Application;
 import javafx.stage.Stage;
 
@@ -9,9 +8,9 @@ import java.io.IOException;
 public class LizardWizardsApplication extends Application {
     @Override
     public void start(Stage stage) throws IOException {
-        stage.setOnCloseRequest(e -> {e.consume(); Utils.closeProgram(stage);});
+        stage.setOnCloseRequest(e -> {e.consume(); ClientUtils.closeProgram(stage);});
         stage.setTitle("Lizard Wizards");
-        Utils.changeScene(stage, 0);
+        ClientUtils.changeScene(stage, 0);
         stage.show();
     }
